fix: move task methods into their own service
