+++ src/main/java/com/interiordesignplanner/room/RoomService.java
@@ -254,78 +254,6 @@ public class RoomService {
     }
 
     /**
-     * Adds Task to checklist for the room
-     * 
-     * 
-     * 
-     * @param task   the task object to be updated
-     * @param roomId retrieves the room object to be updated
-     * @throws RoomNotFoundException if the room is not found
-     * @return the updated room
-     */
-    @Transactional
-    @PreAuthorize("hasRole('DESIGNER')")
-    public RoomDTO addTask(Long roomId, Task task, String username) {
-
-        Room existingRoom = findRoom(roomId);
-
-        findRoomByDesigner(existingRoom, username);
-
-        task.setCompleted(false);
-
-        existingRoom.getChecklist().add(task);
-
-        return roomMapper.toDto(roomRepository.save(existingRoom));
-    }
-
-    /**
-     * Updates task to checklist
-     * 
-     * 
-     * 
-     * @param index      retrieves the item to be updated
-     * @param roomId     retrieves the room object to be updated
-     * @param updateTask the updated task
-     * @throws RoomNotFoundException if the room is not found
-     * @return the updated task is added to the checklist
-     */
-    @Transactional
-    @PreAuthorize("hasRole('DESIGNER')")
-    public RoomDTO editTask(Long roomId, Task updateTask, int index, String username) {
-
-        Room existingRoom = findRoom(roomId);
-
-        findRoomByDesigner(existingRoom, username);
-
-        existingRoom.getChecklist().set(index, updateTask);
-
-        return roomMapper.toDto(roomRepository.save(existingRoom));
-    }
-
-    /**
-     * Delete Task to checklist for the room
-     * 
-     * 
-     * 
-     * @param index  using the task's index to delete
-     * @param roomId retrieves the room object to be updated
-     * @throws RoomNotFoundException if the room is not found
-     * @return the updated room
-     */
-    @Transactional
-    @PreAuthorize("hasRole('DESIGNER')")
-    public void deleteTask(Long roomId, int index, String username) {
-
-        Room existingRoom = findRoom(roomId);
-
-        findRoomByDesigner(existingRoom, username);
-
-        existingRoom.getChecklist().remove(index);
-
-        roomMapper.toDto(roomRepository.save(existingRoom));
-    }
-
-    /**
      * Adds item to inventory
      * 
      *
